historic data bug

+++ src/php/db.php
@@ -68,8 +68,8 @@ function init_finance_database($mysql_conn, $database)
     }
 
     $google_sql = "CREATE TABLE $database.Stock_History(ticker varchar(8) NOT NULL, 
-                   day DATE, PRIMARY KEY (ticker, day), beta float, open int,
-                   high int, low int, close int, volume int)";
+                   day DATE, PRIMARY KEY (ticker, day), beta float, open float,
+                   high float, low float, close float, volume int)";
     if (!mysql_query($google_sql, $mysql_conn))
     {
         print "Error creating historic data table: " . mysql_error() . "\n";
